Expand the "native" target so tests compile and run
