+++ src/include/targets.h
@@ -1,5 +1,7 @@
 #pragma once
+#if !defined TARGET_NATIVE
 #include <Arduino.h>
+#endif
 
 #define UNDEF_PIN (-1)
 
@@ -405,6 +407,9 @@ Designed by NamimnoRC
 /* WS2812 led */
 #define GPIO_PIN_LED_WS2812      PB0
 #define GPIO_PIN_LED_WS2812_FAST PB_0
+#elif defined(TARGET_NATIVE)
+#define IRAM_ATTR
+#include "native.h"
 #else
 #error "Unknown target!"
 #endif
