+++ src/platformio.ini
@@ -19,7 +19,10 @@ extra_configs =
 platform = native
 framework =
 test_ignore = test_embedded
-lib_ignore = BUTTON, CRSF, DAC, EEPROM, FIFO, LQCALC, MSP, OTA, POWERMGNT, SX127xDriver, SX1280Driver
+lib_ignore = BUTTON, DAC, EEPROM, LQCALC, POWERMGNT, SX127xDriver, SX1280Driver
 src_filter = ${common_env_data.src_filter} +<test_main.cpp> -<ESP32*.*> -<STM32*.*> -<ESP8*.*> -<tx_*.cpp> -<rx_*.cpp> -<FHSS.*> -<common.*> -<FastSync.*> -<utils.*> -<config.*>
 build_flags =
+	-std=c++11
+	-Iinclude
+	-D TARGET_NATIVE
 	-D ENABLE_TELEMETRY
